Added deconz
Got mine today from AliE https://www.aliexpress.com/item/4000981707898.html?spm=a2g0s.9042311.0.0.666c4c4duYx5ov
Works perfect with my ConBee 2.
Only ioBroker has an issue to find the plug, but that seems to be a problem of the Adapter, because HASS and the Phosbee App are able to identify and control all 5 (USB + 4 Plugs) entities. The 2 USB ports are controlled by a single entity, but that seems fair to me.