Update GiovanniC_CC2530-CC2592.md
I think the intention is that its one coordinator but can being used as one router. Therefore i think is better putting it the coordinator category (perhaps is possible having more categories but i have not seen one example tagging that) as the manufacturer is writing it is one coordinator in the link.