How to pair with recent firmware
Recently Legrand added capability to easily pair its devices with third-party coordinators. 

Here is a procedure to pair (take from here https://developer.legrand.com/forums/topic/how-to-change-zigbee-channel-for-ref-067773-2)

When you reset the device, it will scan the 16 possible channels (feature available starting NLT-50 firmware). If a network is opened - and if your gateway accepts it -, your device will log on it. If several channels are opened, it could be difficult to detect the correct one. The best is (if possible) not to open several networks, or to use a software with a programing console to choose a specific channel

I have successfully paired my device with ZHA.

To update firmware Legrand coordinator is need. The procedure is described here: https://developer.legrand.com/forums/topic/wireless-switches-not-updating/