Update OWON_THS317-ET.md
I have testen this sensor THS317-ET with zha. According to me and to the documentation it is only a temperature sensor and no humidity sensor. Only the standard model without the external probe has temp and hum.
I have binded it to zha and is was discovered correctly. It seems 100% compatible with zha.