Update iluminize_5128.10.md
Confirmed by me as working with ZHA. Often seen as a light (e.g in Hue app), instructions are available in the manual of the actuator.
Pressing Calib. 2x will calibrate as roller shutter, 4x as tilt shade. Not possible to invert the direction: Treated as inside shade/cover. Works fine when using setposition 0-100%.